pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=42", "wheel"]

[project]
authors = [
  {name = "Vadim Kozyrevskiy"}
]
dependencies = [
  "pydantic==2.*",
  "orjson==3.9.15",
  "decohints==1.0.9",
  "cryptography==42.0.2",
  "stackprinter==0.2.12",
  "redis==5.0.3",
  "asgi-idempotency-header==0.2.0",
  "sqlalchemy[asyncio]",
  "fastapi==0.109.2",
  "fastapi_key_auth==0.15.2",
  "loguru==0.7.2",
  "uvicorn[standard]==0.28.0",
  "sqladmin[full]==0.16.1",
  "httpx==0.27.0"
]
description = ""
name = "fastapi_app"
version = "24.29.01"

[project.optional-dependencies]
git = [
  "pre-commit==3.8.0 "
]
kafka = [
  "aiokafka==0.10.0"
]
lint = [
  "flake8==7.0.0",
  "flake8-pytest",
  "black"
]
otlp = [
  "opentelemetry-instrumentation-fastapi==0.45b0",
  "opentelemetry-instrumentation-redis==0.45b0",
  "opentelemetry-instrumentation-sqlalchemy==0.45b0",
  "opentelemetry-instrumentation-logging==0.45b0",
  "opentelemetry-instrumentation-httpx==0.45b0",
  "opentelemetry-exporter-otlp-proto-http==1.24.0"
]
tests = [
  "pytest~=7.4.2",
  "pytest-asyncio~=0.21.1"
]

[tool.pytest.ini_options]
addopts = "--junit-xml=report.xml"
asyncio_mode = "auto"
junit_family = "xunit1"
testpaths = [
  "tests"
]

[tool.setuptools.packages.find]
where = ["src"]
